+++ src/main/java/org/torproject/collector/conf/SourceType.java
@@ -1,0 +1,11 @@
+/* Copyright 2016 The Tor Project
+ * See LICENSE for licensing information */
+
+package org.torproject.collector.conf;
+
+public enum SourceType {
+  Cache,
+  Local,
+  Remote,
+  Sync;
+}